* cygwin:
  update README regarding cygwin build
  fix, printf's %ld format does not match int64_t on 32bit systems
  fix, avoid warning about unsigned/signed comparison
  fix, nowInMsecs() failed on 32bit systems
  build fix, libxml2 was linked in wrong order
  build fix, include missing <time.h> and <sys/time.h>

+++ src/twsdo.cpp
@@ -459,7 +459,7 @@ void TwsDL::connectTws()
 	int64_t w = nowInMsecs() - lastConnectionTime;
 	if( w < tws_conTimeoutp ) {
 		DEBUG_PRINTF( "Waiting %ldms before connecting again.",
-			(tws_conTimeoutp - w) );
+			(long)(tws_conTimeoutp - w) );
 		curIdleTime = tws_conTimeoutp - w;
 		return;
 	}
